refactor: move trpc code inside trpc folder
