+++ src/trpc/server.ts
@@ -11,8 +11,8 @@ import { type TRPCErrorResponse } from "@trpc/server/rpc";
 import { cookies } from "next/headers";
 import { cache } from "react";
 
-import { appRouter, type AppRouter } from "@/server/api/root";
-import { createTRPCContext } from "@/server/api/trpc";
+import { appRouter, type AppRouter } from "@/trpc/api/root";
+import { createTRPCContext } from "@/trpc/api/trpc";
 import { transformer } from "./shared";
 
 /**
